requirements.txt
groq
python-dotenv
requests
json
streamlit